bot/requirements.txt
discord.py==2.3.2
pymongo==4.7.0
pydantic==2.7.0
python-dotenv==1.0.1
playwright==1.44.0
httpx==0.27.0
